+++ SettMod/SkillStates/BaseStates/deathstate.cs
@@ -1,0 +1,37 @@
+﻿using EntityStates;
+using RoR2;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using UnityEngine;
+
+namespace SettMod.SkillStates.BaseStates
+{
+    class Death : GenericCharacterDeath
+    {
+        public override void OnEnter()
+        {
+            base.OnEnter();
+            Util.PlaySound("Death", base.gameObject);
+            Animator modelAnimator = base.GetModelAnimator();
+            if (modelAnimator)
+            {
+                modelAnimator.CrossFadeInFixedTime("Death", 0.1f);
+            }
+        }
+
+        public override void OnExit()
+        {
+            base.OnExit();
+        }
+
+        public override void FixedUpdate()
+        {
+            base.FixedUpdate();
+        }
+        public override InterruptPriority GetMinimumInterruptPriority()
+        {
+            return InterruptPriority.Death;
+        }
+    }
+}